Improved cljeromq error handling
Trying to cope w/ changed semantics around a recv that returned 0
bytes because the sender sent a NULL vs. one that had nothing to
return because there was no message and we set the :dont-wait
flag.

I'm almost inclined to think this is a bug in the zmq-jni library,
but I may be misunderstanding something.