Debugging my automated protocol builder
Have it up to the point where the client is getting back what should be
the actual preferred protocol version so we can move on to doing
something interesting.

To satisfy spec's tuple, I had to hack up its ::icanhaz request to send
a vector instead of the list that I want.

And the final response is coming in as something that looks like
another stream that I need to take! from, even though I'm pretty sure
that's exactly what the previous step is doing.

Not sure whether the current approach is worth any more effort.

The more I pound my head against this particular wall, the more
convinced I feel that I should just write this as a couple of FSMs
instead.