Fairly serious dependency gutting (WIP)
Started out "just" updating the clojure version. Along with several
other ancient dependencies.

Since I'm trying to update a piece that depends on that to that clojure version.

But at least one of this project's dependencies isn't compatible.

So I wound up pulling that out completely and switching the component
system to integrant. At least in theory. Lots of what's still here is
either totally obsolete or needs to be drastically revamped once I have
a viable replacement for cljeromq.

This version compiles. That's about the only thing I can say about it
with any certainty.